refactor(passes): add list_detect pass
